// src/Framework/SPPC.Tadbir.Web.New/ClientApp/src/app/model/account.ts
// ------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated by a tool.
//     Runtime Version: 1.1.411
//     Template Version: 1.0
//     Generation Date: 09/29/2018 02:10:32 ب.ظ
//  
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.
// </auto-generated>
// ------------------------------------------------------------------------------


export interface Account {
    parentId?: number;
    fiscalPeriodId: number;
    branchId: number;
    companyId: number;
    childCount: number;
    id: number;
    branchScope: number;
    code: string;
    fullCode: string;
    name: string;
    level: number;
    description?: string;
}